refactor(terminal): split bridge adoption lane and dedupe pty table lifecycle
The orphan/adoption feature spanned two oversized modules with parallel
implementations of the same mechanics. No behavior change.

- ptyBridgeChannel.ts: shared newline-JSON feed, frame parser, and exit
  builder used by both bridge transports (owner stdio, control socket)
- ptyBridgeAdoption.ts: control-socket lane moved out of localPtyTerminal,
  which returns to spawn-side only
- ptyHandoff.ts: export/adopt registry orchestration moved out of ptyTable
  behind injected deps; per-entry failure handling unified
- ptyTable.ts: one buildPtyEntry/activateEntry path shared by spawn and
  adoption, so the replay-before-subscription ordering invariant lives in
  exactly one place
- frameStream.ts / ptyEntry.ts: generic stream util and entry record
- localPtyBridge.cjs: protocol version and park budget arrive via spawn
  options instead of hardcoded twins of the contracts constants
