+++ station/src/terminal/pty/localPtyTerminal.ts
@@ -1,7 +1,6 @@
 import { spawn, type ChildProcessWithoutNullStreams } from "node:child_process";
 import { chmodSync, existsSync, statSync } from "node:fs";
 import { createRequire } from "node:module";
-import net from "node:net";
 import os from "node:os";
 import path from "node:path";
 import process from "node:process";
@@ -13,7 +12,6 @@ import type {
   StationTerminalProcess,
   StationTerminalSize,
   StationTerminalSpawnOptions,
-  StationTerminalUnavailable,
 } from "../types.js";
 import {
   CTTY_HELPER_PATH,
@@ -25,12 +23,18 @@ import {
   STATION_CHILD_TERMINAL_ENV,
 } from "./childPtyEnvironment.js";
 import { StationTerminalSpawnError } from "./errors.js";
+import {
+  createJsonLineFeed,
+  parseBridgeLine,
+  type PtyBridgeStreamMessage,
+  toTerminalExit,
+} from "./ptyBridgeChannel.js";
 import { TerminalProcessEmitter } from "./terminalProcessEmitter.js";
 
+export const MIN_COLS = 2;
+export const MIN_ROWS = 1;
 const DEFAULT_COLS = 80;
 const DEFAULT_ROWS = 24;
-const MIN_COLS = 2;
-const MIN_ROWS = 1;
 const BRIDGE_PATH = fileURLToPath(new URL("./localPtyBridge.cjs", import.meta.url));
 let nextTerminalSequence = 0;
 
@@ -40,19 +44,7 @@ type BridgeMessage =
       pid: number;
       bridgePid?: number;
     }
-  | {
-      type: "data";
-      data: string;
-    }
-  | {
-      type: "exit";
-      exitCode: number;
-      signal?: number;
-    }
-  | {
-      type: "error";
-      message: string;
-    };
+  | PtyBridgeStreamMessage;
 
 // node-pty ships its macOS/Linux `spawn-helper` without the execute bit in the
 // bun-extracted layout, and a later `bun install` re-clears it. When it isn't
@@ -130,6 +122,9 @@ export function createLocalPtyTerminal(
   ensureSpawnHelperExecutable();
   const bridgeOptions: Record<string, unknown> = {
     args,
+    // The bridge reports the owner's protocol version in its status reply, so
+    // the contract constant has no hardcoded twin inside the .cjs.
+    bridgeProtocol: PtyBridgeProtocolVersion,
     cols: size.cols,
     command,
     cwd,
@@ -164,10 +159,17 @@ class LocalPtyTerminalProcess implements StationTerminalProcess {
 
   #bridge: ChildProcessWithoutNullStreams;
   #events = new TerminalProcessEmitter();
-  #stdoutBuffer = "";
   #pid: number;
   #bridgePid: number | undefined;
   #size: StationTerminalSize;
+  #feed = createJsonLineFeed((line) => {
+    const message = parseBridgeLine<BridgeMessage>(line, (diagnostic) =>
+      this.#events.emitDiagnostic(diagnostic),
+    );
+    if (message !== undefined) {
+      this.handleBridgeMessage(message);
+    }
+  });
 
   constructor(
     id: string,
@@ -182,38 +184,30 @@ class LocalPtyTerminalProcess implements StationTerminalProcess {
     this.#pid = bridge.pid ?? 0;
 
     bridge.stdout.setEncoding("utf8");
-    bridge.stdout.on("data", (chunk: string) => {
-      this.handleBridgeOutput(chunk);
-    });
+    bridge.stdout.on("data", this.#feed);
     // Bridge stderr is Node diagnostics (warnings, crashes), not terminal
     // output: injected into the data stream it would render as screen content
     // and could corrupt VT parser state mid-escape-sequence.
     bridge.stderr.setEncoding("utf8");
     bridge.stderr.on("data", (chunk: string) => {
-      this.emitDiagnostic(chunk);
+      this.#events.emitDiagnostic(chunk);
     });
     // Writes raced against bridge death (EPIPE) surface here; without a
     // listener they are uncaught exceptions that crash Station.
     bridge.stdin.on("error", (error) => {
-      this.emitDiagnostic(`bridge stdin error: ${error.message}`);
+      this.#events.emitDiagnostic(`bridge stdin error: ${error.message}`);
     });
     bridge.on("error", (error) => {
-      this.emitExit({
-        exitCode: 1,
-      });
-      this.emitData(error.message);
+      this.#events.emitExit({ exitCode: 1 });
+      this.#events.emitData(error.message);
     });
     bridge.on("exit", (code, signal) => {
       // An abnormal bridge death (signal kill, code null) must not read as a
       // clean "exited 0" in the pane title.
       const signalNumber = signal === null ? undefined : signalToNumber(signal);
-      const event: StationTerminalExit = {
-        exitCode: code ?? (signalNumber !== undefined ? 128 + signalNumber : 1),
-      };
-      if (signalNumber !== undefined) {
-        event.signal = signalNumber;
-      }
-      this.emitExit(event);
+      this.#events.emitExit(
+        toTerminalExit(code ?? (signalNumber !== undefined ? 128 + signalNumber : 1), signalNumber),
+      );
     });
   }
 
@@ -242,7 +236,7 @@ class LocalPtyTerminalProcess implements StationTerminalProcess {
   }
 
   write(data: string): void {
-    this.assertActive("write to terminal");
+    this.#events.assertActive("write to terminal");
     // After exit the bridge stdin pipe is dead; a keystroke or forwarded VT
     // query reply must drop silently instead of raising EPIPE.
     if (this.#events.exited) {
@@ -255,7 +249,7 @@ class LocalPtyTerminalProcess implements StationTerminalProcess {
   }
 
   resize(size: StationTerminalSize): void {
-    this.assertActive("resize terminal");
+    this.#events.assertActive("resize terminal");
     if (this.#events.exited) {
       return;
     }
@@ -295,30 +289,6 @@ class LocalPtyTerminalProcess implements StationTerminalProcess {
     }
   }
 
-  private handleBridgeOutput(chunk: string): void {
-    this.#stdoutBuffer += chunk;
-
-    while (true) {
-      const newlineIndex = this.#stdoutBuffer.indexOf("\n");
-      if (newlineIndex === -1) {
-        return;
-      }
-
-      const line = this.#stdoutBuffer.slice(0, newlineIndex);
-      this.#stdoutBuffer = this.#stdoutBuffer.slice(newlineIndex + 1);
-      let message: BridgeMessage;
-      try {
-        message = JSON.parse(line) as BridgeMessage;
-      } catch {
-        // One stray non-JSON line (dependency noise on the bridge's stdout)
-        // must not take the whole pipeline down.
-        this.emitDiagnostic(`unparseable bridge line: ${line.slice(0, 200)}`);
-        continue;
-      }
-      this.handleBridgeMessage(message);
-    }
-  }
-
   private handleBridgeMessage(message: BridgeMessage): void {
     switch (message.type) {
       case "ready":
@@ -326,41 +296,18 @@ class LocalPtyTerminalProcess implements StationTerminalProcess {
         this.#bridgePid = message.bridgePid;
         return;
       case "data":
-        this.emitData(message.data);
+        this.#events.emitData(message.data);
         return;
       case "error":
-        this.emitDiagnostic(`bridge command error: ${message.message}`);
+        this.#events.emitDiagnostic(`bridge command error: ${message.message}`);
         return;
-      case "exit": {
-        const event: StationTerminalExit = {
-          exitCode: message.exitCode,
-        };
-        if (message.signal !== undefined) {
-          event.signal = message.signal;
-        }
-        this.emitExit(event);
-      }
+      case "exit":
+        this.#events.emitExit(toTerminalExit(message.exitCode, message.signal));
     }
-  }
-
-  private emitDiagnostic(message: string): void {
-    this.#events.emitDiagnostic(message);
-  }
-
-  private emitData(data: string): void {
-    this.#events.emitData(data);
-  }
-
-  private emitExit(event: StationTerminalExit): void {
-    this.#events.emitExit(event);
   }
 
   private sendBridgeCommand(command: object): void {
     this.#bridge.stdin.write(`${JSON.stringify(command)}\n`);
-  }
-
-  private assertActive(action: string): void {
-    this.#events.assertActive(action);
   }
 }
 
@@ -371,329 +318,6 @@ export type LocalPtyTerminalRuntime = {
   implementation?: PtyImplementation;
   cttyHelperPath?: string;
 };
-
-type BridgeControlStatus = {
-  bridgeProtocol: number;
-  pid: number;
-  bridgePid: number;
-  cols: number;
-  rows: number;
-  adopted: boolean;
-  exited: boolean;
-  parkedEvicted: boolean;
-  exitCode?: number;
-  signal?: number;
-};
-
-type BridgeControlMessage =
-  | ({ type: "status" } & BridgeControlStatus)
-  | { type: "data"; data: string }
-  | { type: "exit"; exitCode: number; signal?: number }
-  | { type: "error"; message: string; code?: string };
-
-/** A parked bridge rebound as a live terminal; adoption extras ride beside the interface. */
-export type AdoptedPtyBridge = StationTerminalProcess & {
-  /** True when the bridge dropped parked output before adoption. */
-  readonly parkedEvicted: boolean;
-  /** Recorded exit when the PTY died before adoption; undefined while alive. */
-  readonly recordedExit: StationTerminalExit | undefined;
-};
-
-export type AdoptLocalPtyBridgeOptions = {
-  id: string;
-  command: string;
-  controlSocketPath: string;
-  size: StationTerminalSize;
-  /** Fail the adoption when the bridge does not answer in time. */
-  timeoutMs?: number;
-};
-
-const DEFAULT_ADOPT_TIMEOUT_MS = 5_000;
-
-/**
- * ADAPTER
- *
- * Rebinds a parked orphan bridge's control socket as a live terminal process so
- * a new host adopts the surviving PTY instead of respawning its payload.
- * Geometry comes from the bridge's status reply, which is the PTY's truth.
- */
-export function adoptLocalPtyBridge(
-  options: AdoptLocalPtyBridgeOptions,
-): Promise<AdoptedPtyBridge> {
-  return new Promise((resolve, reject) => {
-    const socket = net.connect(options.controlSocketPath);
-    let settled = false;
-    const timer = setTimeout(() => {
-      finishWithError(
-        new Error(`Timed out adopting the parked bridge at ${options.controlSocketPath}.`),
-      );
-    }, options.timeoutMs ?? DEFAULT_ADOPT_TIMEOUT_MS);
-    const finishWithError = (error: Error): void => {
-      if (settled) {
-        return;
-      }
-      settled = true;
-      clearTimeout(timer);
-      socket.destroy();
-      reject(error);
-    };
-    let buffer = "";
-    socket.setEncoding("utf8");
-    socket.on("error", (error) => {
-      finishWithError(error);
-    });
-    socket.on("connect", () => {
-      socket.write(`${JSON.stringify({ type: "adopt" })}\n`);
-    });
-    socket.on("data", (chunk: string) => {
-      buffer += chunk;
-      const newlineIndex = buffer.indexOf("\n");
-      if (newlineIndex === -1) {
-        return;
-      }
-      const line = buffer.slice(0, newlineIndex);
-      const remainder = buffer.slice(newlineIndex + 1);
-      let message: BridgeControlMessage;
-      try {
-        message = JSON.parse(line) as BridgeControlMessage;
-      } catch (error) {
-        finishWithError(
-          error instanceof Error ? error : new Error("Unparseable bridge adoption reply."),
-        );
-        return;
-      }
-      if (message.type === "error") {
-        finishWithError(new Error(message.message));
-        return;
-      }
-      if (message.type !== "status") {
-        finishWithError(new Error(`Unexpected bridge adoption reply ${message.type}.`));
-        return;
-      }
-      if (message.bridgeProtocol !== PtyBridgeProtocolVersion) {
-        finishWithError(
-          new Error(
-            `Bridge control protocol ${message.bridgeProtocol} does not match ${PtyBridgeProtocolVersion}.`,
-          ),
-        );
-        return;
-      }
-      if (settled) {
-        return;
-      }
-      settled = true;
-      clearTimeout(timer);
-      socket.removeAllListeners("error");
-      socket.removeAllListeners("data");
-      resolve(
-        new AdoptedLocalPtyBridgeProcess({
-          id: options.id,
-          command: options.command,
-          size: {
-            cols: normalizeDimension(message.cols, options.size.cols, MIN_COLS),
-            rows: normalizeDimension(message.rows, options.size.rows, MIN_ROWS),
-          },
-          socket,
-          status: message,
-          initialBuffer: remainder,
-        }),
-      );
-    });
-  });
-}
-
-class AdoptedLocalPtyBridgeProcess implements StationTerminalProcess {
-  readonly id: string;
-  readonly command: string;
-  readonly pid: number;
-  readonly bridgePid: number;
-  readonly parkedEvicted: boolean;
-  readonly recordedExit: StationTerminalExit | undefined;
-
-  #socket: net.Socket;
-  #events = new TerminalProcessEmitter();
-  #unavailableListeners = new Set<(event: StationTerminalUnavailable) => void>();
-  #buffer = "";
-  #size: StationTerminalSize;
-  #socketClosed = false;
-  #disposed = false;
-
-  constructor(init: {
-    id: string;
-    command: string;
-    size: StationTerminalSize;
-    socket: net.Socket;
-    status: BridgeControlStatus;
-    initialBuffer: string;
-  }) {
-    const { id, command, size, socket, status, initialBuffer } = init;
-    this.id = id;
-    this.command = command;
-    this.pid = status.pid;
-    this.bridgePid = status.bridgePid;
-    this.parkedEvicted = status.parkedEvicted;
-    if (status.exited) {
-      const event: StationTerminalExit = {
-        exitCode: status.exitCode ?? 1,
-      };
-      if (status.signal !== undefined) {
-        event.signal = status.signal;
-      }
-      this.recordedExit = event;
-    }
-    this.#size = size;
-    this.#socket = socket;
-    socket.setEncoding("utf8");
-    socket.on("data", (chunk: string) => {
-      this.handleChunk(chunk);
-    });
-    socket.on("error", (error) => {
-      this.#events.emitDiagnostic(`adopted bridge socket error: ${error.message}`);
-    });
-    socket.on("close", () => {
-      this.handleSocketClose();
-    });
-    if (initialBuffer.length > 0) {
-      this.handleChunk(initialBuffer);
-    }
-  }
-
-  get size(): StationTerminalSize {
-    return this.#size;
-  }
-
-  onData(listener: (data: string) => void): StationTerminalDisposable {
-    return this.#events.onData(listener);
-  }
-
-  onExit(listener: (event: StationTerminalExit) => void): StationTerminalDisposable {
-    return this.#events.onExit(listener);
-  }
-
-  onDiagnostic(listener: (message: string) => void): StationTerminalDisposable {
-    return this.#events.onDiagnostic(listener);
-  }
-
-  onUnavailable(
-    listener: (event: StationTerminalUnavailable) => void,
-  ): StationTerminalDisposable {
-    this.#events.assertActive("subscribe to terminal unavailability");
-    this.#unavailableListeners.add(listener);
-    return {
-      dispose: () => {
-        this.#unavailableListeners.delete(listener);
-      },
-    };
-  }
-
-  write(data: string): void {
-    this.#events.assertActive("write to terminal");
-    if (this.#events.exited || this.#socketClosed) {
-      return;
-    }
-    this.sendControl({ type: "write", data });
-  }
-
-  resize(size: StationTerminalSize): void {
-    this.#events.assertActive("resize terminal");
-    if (this.#events.exited || this.#socketClosed) {
-      return;
-    }
-    this.#size = normalizeSize(size);
-    this.sendControl({ type: "resize", cols: this.#size.cols, rows: this.#size.rows });
-  }
-
-  kill(signal?: string): void {
-    if (this.#disposed || this.#events.exited || this.#socketClosed) {
-      return;
-    }
-    this.sendControl({ type: "kill", signal });
-  }
-
-  dispose(): void {
-    if (this.#disposed) {
-      return;
-    }
-    this.#disposed = true;
-    const exited = this.#events.exited;
-    this.#events.dispose();
-    this.#unavailableListeners.clear();
-    if (!exited && !this.#socketClosed) {
-      // Kill before ending the socket: an EOF without a kill tells the bridge
-      // its adopter died, which re-parks the PTY instead of tearing it down.
-      this.sendControl({ type: "kill" });
-    }
-    this.#socket.end();
-  }
-
-  private handleChunk(chunk: string): void {
-    this.#buffer += chunk;
-    while (true) {
-      const newlineIndex = this.#buffer.indexOf("\n");
-      if (newlineIndex === -1) {
-        return;
-      }
-      const line = this.#buffer.slice(0, newlineIndex);
-      this.#buffer = this.#buffer.slice(newlineIndex + 1);
-      let message: BridgeControlMessage;
-      try {
-        message = JSON.parse(line) as BridgeControlMessage;
-      } catch {
-        this.#events.emitDiagnostic(`unparseable adopted-bridge line: ${line.slice(0, 200)}`);
-        continue;
-      }
-      this.handleControlMessage(message);
-    }
-  }
-
-  private handleControlMessage(message: BridgeControlMessage): void {
-    switch (message.type) {
-      case "data":
-        this.#events.emitData(message.data);
-        return;
-      case "exit": {
-        const event: StationTerminalExit = {
-          exitCode: message.exitCode,
-        };
-        if (message.signal !== undefined) {
-          event.signal = message.signal;
-        }
-        this.#events.emitExit(event);
-        return;
-      }
-      case "error":
-        this.#events.emitDiagnostic(`adopted bridge command error: ${message.message}`);
-        return;
-      case "status":
-        // A status frame after adoption is a stray probe reply; nothing to do.
-        return;
-    }
-  }
-
-  private handleSocketClose(): void {
-    this.#socketClosed = true;
-    if (this.#disposed || this.#events.exited) {
-      return;
-    }
-    // The bridge vanished or re-parked without an exit frame: attachment is
-    // lost, but no exit evidence exists, so report unavailability, not death.
-    this.#events.emitDiagnostic("adopted bridge socket closed before terminal exit.");
-    for (const listener of [...this.#unavailableListeners]) {
-      listener({
-        code: "ADOPTED_BRIDGE_LOST",
-        message: "The adopted bridge connection closed before the terminal exited.",
-      });
-    }
-  }
-
-  private sendControl(command: object): void {
-    this.#socket.write(`${JSON.stringify(command)}\n`, (error) => {
-      if (error !== undefined && error !== null) {
-        this.#events.emitDiagnostic(`adopted bridge send failed: ${error.message}`);
-      }
-    });
-  }
-}
 
 export function resolvePtyImplementation(
   value: string | undefined,
@@ -743,14 +367,19 @@ export function defaultShellArgs(): string[] {
   return ["-i"];
 }
 
-function normalizeSize(size: Partial<StationTerminalSize> | undefined): StationTerminalSize {
+/** Geometry floor shared with the adoption lane, which clamps bridge-reported sizes the same way. */
+export function normalizeSize(size: Partial<StationTerminalSize> | undefined): StationTerminalSize {
   return {
     cols: normalizeDimension(size?.cols, DEFAULT_COLS, MIN_COLS),
     rows: normalizeDimension(size?.rows, DEFAULT_ROWS, MIN_ROWS),
   };
 }
 
-function normalizeDimension(value: number | undefined, fallback: number, minimum: number): number {
+export function normalizeDimension(
+  value: number | undefined,
+  fallback: number,
+  minimum: number,
+): number {
   if (value === undefined) {
     return fallback;
   }
